requirements.txt
langchain-chroma
langchain_community
langchain_openai
langchain_text_splitters

python-dotenv